% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-wave-channel.R
\name{add_crop_lines}
\alias{add_crop_lines}
\title{Add vertical lines to plotly object that can be used for cropping}
\usage{
add_crop_lines(pl_plotly, audio_params)
}
\arguments{
\item{pl_plotly}{a \code{plotly} object}

\item{audio_params}{List of audio parameters returned from \code{process_wave_channel}.}
}
\description{
Add vertical lines to plotly object that can be used for cropping
}
\keyword{internal}
