% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-wave-channel.R
\name{add_play_tracker_line}
\alias{add_play_tracker_line}
\title{Add vertical line to plotly object to track current play time}
\usage{
add_play_tracker_line(pl_plotly, color = "red")
}
\arguments{
\item{pl_plotly}{a \code{plotly} object}

\item{color}{color of line}
}
\description{
Add vertical line to plotly object to track current play time
}
\keyword{internal}
