% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-mp3.R
\name{adjust_audio_speed}
\alias{adjust_audio_speed}
\title{Speed up or slow down mp3 file}
\usage{
adjust_audio_speed(audio_obj, speed_factor = 2)
}
\arguments{
\item{audio_obj}{a \code{tuneR} audio object}

\item{speed_factor}{multiplier of song frequency}
}
\description{
Speed up or slow down mp3 file
}
\keyword{internal}
