% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-audio-playback.R
\name{audio_playpack_ui}
\alias{audio_playpack_ui}
\alias{audio_playpack_server}
\title{Create shiny module for altering the playback of an audio track}
\usage{
audio_playpack_ui(id)

audio_playpack_server(id, audio_choices, audio_dir, audio_select)
}
\description{
Create shiny module for altering the playback of an audio track
}
\section{Functions}{
\itemize{
\item \code{audio_playpack_ui()}: Create module UI for altering the playback of an audio track

\item \code{audio_playpack_server()}: Create module server for altering the playback of an audio track

}}
\keyword{internal}
