% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{check_audio_py}
\alias{check_audio_py}
\title{Do something with python}
\usage{
check_audio_py(audio_path = file.path(EXAMPLE_AUDIO_DIR, "flowers.mp3"))
}
\arguments{
\item{audio_path}{path to audio file}
}
\description{
Example with working with \code{audio_path} and importing a python package that
\emph{comes with python}
}
\details{
The \code{os} python module comes with python:
\itemize{
\item We dont need a virtual environment for those packages either (much quicker)
\item Virtual environment (or conda) is only needed when we have to \emph{install}
packages.
}
}
\keyword{internal}
