% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-mp3.R
\name{crop_mp3}
\alias{crop_mp3}
\title{Crop mp3 file to specified limits}
\usage{
crop_mp3(audio_obj, limits = c(0, 10))
}
\arguments{
\item{audio_obj}{a \code{tuneR} audio object}

\item{limits}{vector specifying the start and end time}
}
\description{
Crop mp3 file to specified limits
}
\keyword{internal}
