% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dev_bug}
\alias{dev_bug}
\title{Format error message that shouldn't happen}
\usage{
dev_bug(error_msg)
}
\arguments{
\item{error_msg}{text to error with}
}
\description{
Format error message that shouldn't happen
}
\keyword{internal}
