% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dev_warning}
\alias{dev_warning}
\title{Format warning message that shouldn't happen}
\usage{
dev_warning(warn_msg)
}
\arguments{
\item{warn_msg}{text to warn with}
}
\description{
Format warning message that shouldn't happen
}
\keyword{internal}
