% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-download-audio.R
\name{download_audio_ui}
\alias{download_audio_ui}
\alias{download_audio_server}
\title{Create shiny module for saving youtube videos as audio files}
\usage{
download_audio_ui(id)

download_audio_server(id)
}
\description{
Create shiny module for saving youtube videos as audio files
}
\section{Functions}{
\itemize{
\item \code{download_audio_ui()}: Create module UI for saving youtube videos as audio files

\item \code{download_audio_server()}: Create module server for saving youtube videos as audio files

}}
\keyword{internal}
