% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-equalizer.R
\name{eq_smooth_line}
\alias{eq_smooth_line}
\title{Create smoothed line}
\usage{
eq_smooth_line(data_pl, xvar = "index", nsteps = 50, jitter = 100)
}
\arguments{
\item{data_pl}{plot data}

\item{xvar}{x variable to perform \code{lm()} with (y is always \code{shift}).}

\item{nsteps}{number of steps to make in between points}

\item{jitter}{amount to jitter values to increase fake variability ribbon}
}
\description{
Create smoothed line
}
\keyword{internal}
