% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{EXAMPLE_AUDIO_DIR}
\alias{EXAMPLE_AUDIO_DIR}
\title{Example audio directory}
\format{
An object of class \code{character} of length 1.
}
\usage{
EXAMPLE_AUDIO_DIR
}
\description{
Example audio directory
}
\keyword{datasets}
