% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{example_py_func}
\alias{example_py_func}
\title{Simple python function}
\usage{
example_py_func(x = 5, y = 10)
}
\arguments{
\item{x}{a number}

\item{y}{a number}
}
\description{
Simple python function
}
\details{
This function does not require imports or virtual environment (much faster)
}
