% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_seconds}
\alias{format_seconds}
\title{Format seconds as \code{{minutes:seconds}}}
\usage{
format_seconds(time, as_date = TRUE)
}
\arguments{
\item{time}{time in seconds}
}
\value{
character string formatted as \code{{minutes:seconds}}
}
\description{
Format seconds as \code{{minutes:seconds}}
}
\examples{

format_seconds(80, as_date = FALSE)

}
\keyword{internal}
