% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-mp3.R
\name{get_audio_dur}
\alias{get_audio_dur}
\title{Get duration of audio file}
\usage{
get_audio_dur(audio_obj)
}
\arguments{
\item{audio_obj}{a \code{tuneR} audio object}
}
\value{
a vector of \code{c(0, duration)}
}
\description{
Get duration of audio file
}
\keyword{internal}
