% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howler.R
\name{howlerSeekSlider}
\alias{howlerSeekSlider}
\title{Seek Slider}
\usage{
howlerSeekSlider(id)
}
\arguments{
\item{id}{ID given to the volume slider. For it to work with the \code{howler}, the ID
must match that of the \code{howler}.}
}
\value{
A slider element of class \code{howler-seek-slider} that will display the position of the current track
playing.
}
\description{
A UI element that can be included with a \code{\link{howler}} to manually change the location of the track.
}
\examples{
 \dontrun{
if (interactive()) {
  library(shiny)
  sound_file <- file.path("audio", list.files(EXAMPLE_AUDIO_DIR))[1]
  addResourcePath("audio", EXAMPLE_AUDIO_DIR)

  ui <- fluidPage(
    title = "howler.js Player",
    howler(elementId = "sound", sound_file),
    howlerPlayPauseButton("sound"),
    howlerSeekSlider("sound")
  )

  server <- function(input, output, session) {
  }

  shinyApp(ui, server)
}
}

}
\keyword{internal}
