% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-import-audio.R
\name{import_audio_ui}
\alias{import_audio_ui}
\alias{import_audio_server}
\title{Create shiny module for choosing an audio directory and selecting which files to import}
\usage{
import_audio_ui(
  id,
  title = "Select a folder containing audio files",
  box_title = "Set Audio Library",
  collapsible = FALSE,
  collapsed = FALSE,
  status = "primary"
)

import_audio_server(id, audio_dir = NULL, title = "Starting Directory")
}
\arguments{
\item{id}{module id}

\item{title}{label fed to \code{shinyDirButton}}

\item{box_title}{label of the first \code{shinydashboardPlus::box}}

\item{collapsible}{whether the module can be collapsed/hidden}

\item{collapsed}{whether the module should start off as collapsed}

\item{status}{the color of the box. See \code{shinydashboardPlus::box} status}

\item{audio_dir}{an initial audio directory}
}
\description{
Create shiny module for choosing an audio directory and selecting which files to import
}
\section{Functions}{
\itemize{
\item \code{import_audio_ui()}: Create module UI for choosing an audio directory
and selecting which files to import

\item \code{import_audio_server()}: Create module server for choosing an audio
directory and selecting which files to import

}}
\keyword{internal}
