% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-python-pkgs.R
\name{install_py_pkgs}
\alias{install_py_pkgs}
\title{Install python packages to an environment}
\usage{
install_py_pkgs(
  py_pkgs = c("scipy", "pandas"),
  virtual_env = FALSE,
  env_name = SONGCLIP_PYTHON_ENV,
  update = FALSE,
  python_version = NULL
)
}
\arguments{
\item{py_pkgs}{vector of python packages to install}

\item{virtual_env}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, use a virtual environment.}

\item{env_name}{a virtual environment name}

\item{update}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, update packages that are
already installed.}

\item{python_version}{The requested Python version. Ignored when attempting
to install with a Python virtual environment.}
}
\description{
Install python packages to an environment
}
\details{
see \code{?reticulate::py_install} for more details
}
\keyword{internal}
