% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-equalizer.R
\name{make_equalizer_data}
\alias{make_equalizer_data}
\title{Make equalizer data}
\usage{
make_equalizer_data(starting_vals = rep(0, 6))
}
\arguments{
\item{starting_vals}{vector of starting values}
}
\description{
Make equalizer data
}
