% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-make-equalizer.R
\name{make_equalizer_ui}
\alias{make_equalizer_ui}
\alias{make_equalizer_server}
\title{Create shiny module for modifying the equalizer}
\usage{
make_equalizer_ui(id)

make_equalizer_server(id)
}
\description{
Create shiny module for modifying the equalizer
}
\section{Functions}{
\itemize{
\item \code{make_equalizer_ui()}: Create module UI for modifying the equalizer

\item \code{make_equalizer_server()}: Create module server for modifying the equalizer

}}
\keyword{internal}
