% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howler.R
\name{make_howler_ui}
\alias{make_howler_ui}
\title{Make howler UI for use in \code{\link{audio_playpack_ui}}}
\usage{
make_howler_ui(
  files,
  howler_id = "howler",
  seek_ping_rate = 100,
  include_current_track = TRUE,
  width = "300px",
  options = list(),
  ...
)
}
\arguments{
\item{files}{Files that will be used in the player.}

\item{howler_id}{ID to correspond to the howler object and its related buttons.}

\item{seek_ping_rate}{Number of milliseconds between each update of \verb{input$\{id\}_seek} while playing. Default is
set to 1000. If set to 0, then \verb{input$\{id\}_seek} will not exist. Decreasing this number too much could increase
the cross-talk between the server and client to the point where significant lags occur.}

\item{include_current_track}{Logical, should the current track be included in the UI of the module?}

\item{width}{Width (in pixels) of the player. Defaults to 300px.}
}
\description{
Make howler UI for use in \code{\link{audio_playpack_ui}}
}
\keyword{internal}
