% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-mp3.R
\name{mp3_speed}
\alias{mp3_speed}
\title{Speed up or slow down mp3 file}
\usage{
mp3_speed(audio_obj, speed = 2)
}
\arguments{
\item{audio_obj}{a \code{tuneR} audio object}

\item{speed}{multiplier of song frequency}
}
\description{
Speed up or slow down mp3 file
}
\keyword{internal}
