% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{no_margin_row}
\alias{no_margin_row}
\title{Substitute for \code{\link{fluidRow}} with no margins}
\usage{
no_margin_row(..., style = NULL, color = NULL, bg_color = NULL)
}
\arguments{
\item{...}{HTML tags}

\item{style}{html formatted string to append to the \code{style} argument of a \code{fluidRow}.}

\item{color}{font color of the \code{fluidRow}.}

\item{bg_color}{background color of the \code{fluidRow}.}
}
\description{
Plotly behaves a little funky in \code{fluidRows}'s. Manually removing the left
and right margins was the easiest method for removing the artificial horizontal
scrollbars that get added.
}
\keyword{internal}
