% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-mp3.R
\name{plot_mp3}
\alias{plot_mp3}
\title{Plot the audio bins (looks cool)}
\usage{
plot_mp3(mp3_file)
}
\arguments{
\item{mp3_file}{file path to an MP3 file}
}
\description{
Plot the audio bins (looks cool)
}
\details{
This takes a while for long videos
}
