% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-wave-channel.R
\name{plot_wave_channel_base}
\alias{plot_wave_channel_base}
\title{Original plotting method}
\usage{
plot_wave_channel_base(
  audio_data,
  audio_params,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  plot_title = NULL,
  axes = TRUE,
  center = TRUE
)
}
\arguments{
\item{audio_data}{Audio data returned from \code{process_wave_channel}.}

\item{audio_params}{List of audio parameters returned from \code{process_wave_channel}.}

\item{ylim}{y-limit. Will be determined if not set.}

\item{xlab}{label for x-axis. Defaults to \code{xunit} if not set.}

\item{ylab}{label for y-axis. Will be determined if not set.}

\item{plot_title}{a title for the plot (optional).}

\item{axes}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, add axes to the plot}

\item{center}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, center the plot}
}
\description{
Original plotting method
}
\keyword{internal}
