% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-wave-channel.R
\name{plot_wave_channel_fancy}
\alias{plot_wave_channel_fancy}
\title{New plotting method}
\usage{
plot_wave_channel_fancy(
  audio_data,
  audio_params,
  source = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  plot_title = NULL,
  axes = TRUE,
  hollow = FALSE,
  show_y_axis = FALSE,
  line_color = "#ADD8E6",
  ft_color = "lightgrey",
  bg_color = "#252525"
)
}
\arguments{
\item{audio_data}{Audio data returned from \code{process_wave_channel}.}

\item{audio_params}{List of audio parameters returned from \code{process_wave_channel}.}

\item{ylim}{y-limit. Will be determined if not set.}

\item{xlab}{label for x-axis. Defaults to \code{xunit} if not set.}

\item{ylab}{label for y-axis. Will be determined if not set.}

\item{plot_title}{a title for the plot (optional).}

\item{axes}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, add axes to the plot}

\item{hollow}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, make the plot hollow}
}
\description{
New plotting method
}
\keyword{internal}
