% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{process_audio_py}
\alias{process_audio_py}
\title{Do something with python}
\usage{
process_audio_py(
  audio_path = file.path(EXAMPLE_AUDIO_DIR, "flowers.mp3"),
  setup_env = FALSE
)
}
\arguments{
\item{audio_path}{path to audio file}
}
\description{
Example with working with \code{audio_path}
}
\details{
The \code{os} python module comes with python:
\itemize{
\item We apparently dont need a virtual environment for those packages either
\item Virtual environment (or conda) is only needed when we have to \emph{install}
packages.
}
}
\keyword{internal}
