% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-wave-channel.R
\name{process_wave_channel}
\alias{process_wave_channel}
\title{Process Wave channel}
\usage{
process_wave_channel(audio_obj, xunit = "Time", simplify = TRUE, nr = 2500)
}
\arguments{
\item{audio_obj}{a \code{tuneR} audio object.}

\item{xunit}{unit for x-axis. One of \code{c("Time", "Samples")}.}

\item{simplify}{Logical (\code{TRUE}/\code{FALSE}). Whether to simplify large audio files. Defaults to \code{TRUE}.}

\item{nr}{noise reduction. Only takes affect if \code{simplify = TRUE}.}
}
\description{
Process Wave channel
}
\keyword{internal}
