% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-import-audio.R
\name{set_root_opts}
\alias{set_root_opts}
\title{Set root directories for use with \code{shinyDirChoose}}
\usage{
set_root_opts(audio_dir, title)
}
\arguments{
\item{audio_dir}{an initial audio directory}

\item{title}{label of root directory for \code{audio_dir}}
}
\description{
Set root directories for use with \code{shinyDirChoose}
}
\keyword{internal}
