% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-python-pkgs.R
\name{setup_py_env}
\alias{setup_py_env}
\title{Set up virtual python environment with required packages}
\usage{
setup_py_env(py_pkgs = c("scipy", "pandas"), env = SONGCLIP_PYTHON_ENV)
}
\arguments{
\item{py_pkgs}{vector of python packages to install}

\item{env}{a virtual environment name}
}
\description{
Set up virtual python environment with required packages
}
\keyword{internal}
