% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-audio-playback.R
\name{shift_selected_song}
\alias{shift_selected_song}
\title{Circular shift the order of the songs based on which one is selected}
\usage{
shift_selected_song(audio_select, audio_choices)
}
\arguments{
\item{audio_select}{the \code{basename} of the file selected}

\item{audio_choices}{vector of file paths}
}
\description{
Circular shift the order of the songs based on which one is selected
}
