% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-python-pkgs.R
\name{shutdown_virtual_env}
\alias{shutdown_virtual_env}
\title{Shut down virtual environment}
\usage{
shutdown_virtual_env(env = SONGCLIP_PYTHON_ENV, force = TRUE)
}
\arguments{
\item{env}{a virtual environment name}

\item{force}{Logical (\code{TRUE}/\code{FALSE}). If \code{FALSE}, confirm before removing
packages or virtual environments}
}
\description{
Shut down virtual environment
}
\details{
Potentially call this at some point

This will shut down the environment...
Shutting it down means you'd have to install the packages again, should we
need them again. Given that, we should only shut this down -if we are done-
using the python packages.

i.e. probably dont run this once per function, but after all python related functions
are done (probably when the app closes, or maybe not at all (i.e. it runs for the
whole session))
}
\keyword{internal}
