% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip-song.R
\docType{data}
\name{song_clip_theme}
\alias{song_clip_theme}
\title{Set theme for songClip app}
\format{
An object of class \code{css} (inherits from \code{html}, \code{character}) of length 1.
}
\usage{
song_clip_theme
}
\description{
Set theme for songClip app
}
\keyword{internal}
