% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\docType{data}
\name{SONGCLIP_PYTHON_DIR}
\alias{SONGCLIP_PYTHON_DIR}
\title{Package directory containing python scripts}
\format{
An object of class \code{character} of length 1.
}
\usage{
SONGCLIP_PYTHON_DIR
}
\description{
Package directory containing python scripts
}
\keyword{datasets}
