% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-python-pkgs.R
\docType{data}
\name{SONGCLIP_PYTHON_ENV}
\alias{SONGCLIP_PYTHON_ENV}
\title{Default environment name for python packages}
\format{
An object of class \code{character} of length 1.
}
\usage{
SONGCLIP_PYTHON_ENV
}
\description{
Can be use for virtual of conda environment
}
\keyword{datasets}
