% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-audio-playback.R
\name{tracker_js}
\alias{tracker_js}
\title{This is how we will update the plot line; we \emph{may} need another method for attaining
the seek information/storing it on the client side}
\usage{
tracker_js(plot_id = "audio_plot", ns = "module_id")
}
\description{
This is how we will update the plot line; we \emph{may} need another method for attaining
the seek information/storing it on the client side
}
