% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-equalizer.R
\name{update_equalizer_data}
\alias{update_equalizer_data}
\title{Update equalizer data based on values in sliders}
\usage{
update_equalizer_data(eq_data, slider_updates)
}
\arguments{
\item{eq_data}{dataset in the form returned by \code{make_equalizer_data}}

\item{slider_updates}{table of values returned by vertical sliders}
}
\description{
Update equalizer data based on values in sliders
}
