% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-audio-playback.R,
%   R/module-download-audio.R
\name{audio_playpack_ui}
\alias{audio_playpack_ui}
\alias{audio_playpack_server}
\alias{download_audio_ui}
\alias{download_audio_server}
\title{Create shiny module for saving youtube videos as audio files}
\usage{
audio_playpack_ui(id)

audio_playpack_server(id, audio_choices, audio_dir, audio_select)

download_audio_ui(id)

download_audio_server(id)
}
\description{
Create shiny module for saving youtube videos as audio files

Create shiny module for saving youtube videos as audio files
}
\section{Functions}{
\itemize{
\item \code{audio_playpack_ui()}: Create module UI for saving youtube videos as audio files

\item \code{audio_playpack_server()}: Create module server for saving youtube videos as audio files

\item \code{download_audio_ui()}: Create module UI for saving youtube videos as audio files

\item \code{download_audio_server()}: Create module server for saving youtube videos as audio files

}}
\keyword{internal}
