% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-python-pkgs.R
\name{import_py_pkgs}
\alias{import_py_pkgs}
\title{Import python packages into the environment}
\usage{
import_py_pkgs(py_pkgs = c("scipy", "pandas"))
}
\arguments{
\item{py_pkgs}{vector of python packages to install}
}
\description{
Import python packages into the environment
}
