% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-equalizer.R
\name{make_equalizer_plot}
\alias{make_equalizer_plot}
\title{Make Equalizer plot}
\usage{
make_equalizer_plot(shift_bounds = c(-12, 12))
}
\description{
TODO: add mechanism for moving the points (drag and drop)
}
\details{
potential resource: https://stackoverflow.com/questions/47280032/draggable-line-chart-in-r-shiny
}
\section{worst case can use sliders or something, but drag and drop is much more}{
user friendly.
}

