% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-equalizer.R
\name{make_equalizer_plot}
\alias{make_equalizer_plot}
\title{Make Equalizer plot}
\usage{
make_equalizer_plot(eq_data = make_equalizer_data(), shift_bounds = c(-12, 12))
}
\arguments{
\item{eq_data}{equalizer data with expected columns returned from
\code{make_equalizer_data()}}

\item{shift_bounds}{y-axis limits.}
}
\value{
\code{plotly} object with draggable points
}
\description{
Make Equalizer plot
}
\examples{

\dontrun{
data_pl <- make_equalizer_data(round(jitter(c(rep(0, 2), 2, rep(0, 3)), 4)))
pl_plotly <- make_equalizer_plot(data_pl)
pl_plotly
}

}
\keyword{internal}
