% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-equalizer.R
\name{stagger_eq_ribbon}
\alias{stagger_eq_ribbon}
\title{Add geom_ribbon to create shadow effect}
\usage{
stagger_eq_ribbon(pl, data_pl, ntimes = 12, color = "#0BDA51", alpha = 0.05)
}
\arguments{
\item{data_pl}{plot data}

\item{ntimes}{number of ribbons to make}

\item{color}{string. Color of ribbon fill.}

\item{alpha}{numeric. Value of alpha, passed to \code{ggplot::geom_ribbon()}}
}
\description{
Add geom_ribbon to create shadow effect
}
\keyword{internal}
