% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-tune-audio.R
\name{tune_audio_ui}
\alias{tune_audio_ui}
\alias{tune_audio_server}
\title{Create shiny module for tuning an audio file}
\usage{
tune_audio_ui(id)

tune_audio_server(id, audio_files)
}
\arguments{
\item{audio_files}{reactive vector of audio files}
}
\description{
Create shiny module for tuning an audio file
}
\section{Functions}{
\itemize{
\item \code{tune_audio_ui()}: Create module UI for tuning an audio file

\item \code{tune_audio_server()}: Create module server for tuning an audio file

}}
\keyword{internal}
