% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-make-equalizer.R
\name{update_equalizer_data}
\alias{update_equalizer_data}
\title{Update equalizer data in response to changes in plotly shape anchors}
\usage{
update_equalizer_data(eq_data, event_data)
}
\arguments{
\item{eq_data}{dataset in the form returned by \code{make_equalizer_data}}

\item{event_data}{data returned by \code{plotly::event_data()}}
}
\description{
Update equalizer data in response to changes in plotly shape anchors
}
\keyword{internal}
